feat(logging): change system info crate
